refactor: modularize setup script by extracting service-specific configurations into dedicated files.
