+++ scripts/setup/setup.ts
@@ -13,17 +13,15 @@ import {
 	initializeEnvFile as initializeEnvFileInternal,
 } from "./envFileManager";
 import { promptConfirm, promptInput, promptList } from "./promptHelpers";
+import { administratorEmail } from "./services/administratorSetup";
+import { apiSetup } from "./services/apiSetup";
+import { caddySetup } from "./services/caddySetup";
+import { setCI } from "./services/ciSetup";
+import { cloudbeaverSetup } from "./services/cloudbeaverSetup";
+import { minioSetup } from "./services/minioSetup";
+import { postgresSetup } from "./services/postgresSetup";
 import { updateEnvVariable } from "./updateEnvVariable";
-import {
-	generateJwtSecret,
-	validateCloudBeaverAdmin,
-	validateCloudBeaverPassword,
-	validateCloudBeaverURL,
-	validateEmail,
-	validatePort,
-	validatePositiveInteger,
-	validateURL,
-} from "./validators";
+import { validatePositiveInteger } from "./validators";
 
 // Re-export validators for backward compatibility
 export {
@@ -480,7 +478,7 @@ export async function metricsSetup(
 	}
 	return answers;
 }
-async function handlePromptError(err: unknown): Promise<never> {
+export async function handlePromptError(err: unknown): Promise<never> {
 	console.error(err);
 	if (backupCreated) {
 		try {
@@ -519,29 +517,10 @@ export async function initializeEnvFile(answers: SetupAnswers): Promise<void> {
 		throw error;
 	}
 }
-export async function setCI(answers: SetupAnswers): Promise<SetupAnswers> {
-	try {
-		answers.CI = await promptList("CI", "Set CI:", ["true", "false"], "false");
-	} catch (err) {
-		await handlePromptError(err);
-	}
-	return answers;
-}
-export async function administratorEmail(
-	answers: SetupAnswers,
-): Promise<SetupAnswers> {
-	try {
-		answers.API_ADMINISTRATOR_USER_EMAIL_ADDRESS = await promptInput(
-			"API_ADMINISTRATOR_USER_EMAIL_ADDRESS",
-			"Enter administrator user email address:",
-			"administrator@email.com",
-			validateEmail,
-		);
-	} catch (err) {
-		await handlePromptError(err);
-	}
-	return answers;
-}
+// These functions are now exported from their respective modules in services/
+// and are kept here for backward compatibility if needed, or renamed to avoid conflict.
+// However, the internal implementation will be removed to avoid duplication.
+
 export async function reCaptchaSetup(
 	answers: SetupAnswers,
 ): Promise<SetupAnswers> {
@@ -738,450 +717,6 @@ export async function oauthSetup(answers: SetupAnswers): Promise<SetupAnswers> {
 		}
 
 		console.log("\nOAuth provider configuration completed!");
-	} catch (err) {
-		await handlePromptError(err);
-	}
-	return answers;
-}
-
-export async function apiSetup(answers: SetupAnswers): Promise<SetupAnswers> {
-	try {
-		answers.API_BASE_URL = await promptInput(
-			"API_BASE_URL",
-			"API base URL:",
-			"http://127.0.0.1:4000",
-			validateURL,
-		);
-		answers.API_HOST = await promptInput("API_HOST", "API host:", "0.0.0.0");
-		answers.API_PORT = await promptInput(
-			"API_PORT",
-			"API port:",
-			"4000",
-			validatePort,
-		);
-		answers.API_IS_APPLY_DRIZZLE_MIGRATIONS = await promptList(
-			"API_IS_APPLY_DRIZZLE_MIGRATIONS",
-			"Apply Drizzle migrations?",
-			["true", "false"],
-			"true",
-		);
-		answers.API_IS_GRAPHIQL = await promptList(
-			"API_IS_GRAPHIQL",
-			"Enable GraphQL?",
-			["true", "false"],
-			answers.CI === "false" ? "true" : "false",
-		);
-		answers.API_IS_PINO_PRETTY = await promptList(
-			"API_IS_PINO_PRETTY",
-			"Enable Pino Pretty logs?",
-			["true", "false"],
-			answers.CI === "false" ? "true" : "false",
-		);
-		answers.API_JWT_EXPIRES_IN = await promptInput(
-			"API_JWT_EXPIRES_IN",
-			"JWT expiration (ms):",
-			"2592000000",
-		);
-		const jwtSecret = generateJwtSecret();
-		answers.API_JWT_SECRET = await promptInput(
-			"API_JWT_SECRET",
-			"JWT secret:",
-			jwtSecret,
-			(input: string) => {
-				const trimmed = input.trim();
-				if (trimmed.length < 128) {
-					return "JWT secret must be at least 128 characters long.";
-				}
-				return true;
-			},
-		);
-
-		answers.API_EMAIL_VERIFICATION_TOKEN_EXPIRES_SECONDS = await promptInput(
-			"API_EMAIL_VERIFICATION_TOKEN_EXPIRES_SECONDS",
-			"Email verification token expiration (seconds):",
-			"86400",
-			(input: string) => {
-				const seconds = Number.parseInt(input, 10);
-				if (Number.isNaN(seconds) || seconds < 60) {
-					return "Expiration must be at least 60 seconds.";
-				}
-				return true;
-			},
-		);
-
-		const emailVerificationSecret = generateJwtSecret();
-		answers.API_EMAIL_VERIFICATION_TOKEN_HMAC_SECRET = await promptInput(
-			"API_EMAIL_VERIFICATION_TOKEN_HMAC_SECRET",
-			"Email verification HMAC secret:",
-			emailVerificationSecret,
-			(input: string) => {
-				const trimmed = input.trim();
-				if (trimmed.length < 32) {
-					return "HMAC secret must be at least 32 characters long.";
-				}
-				return true;
-			},
-		);
-
-		answers.API_LOG_LEVEL = await promptList(
-			"API_LOG_LEVEL",
-			"Log level:",
-			["info", "debug"],
-			answers.CI === "true" ? "info" : "debug",
-		);
-		answers.API_MINIO_ACCESS_KEY = await promptInput(
-			"API_MINIO_ACCESS_KEY",
-			"Minio access key:",
-			"talawa",
-		);
-		answers.API_MINIO_END_POINT = await promptInput(
-			"API_MINIO_END_POINT",
-			"Minio endpoint:",
-			"minio",
-		);
-		answers.API_MINIO_PORT = await promptInput(
-			"API_MINIO_PORT",
-			"Minio port:",
-			"9000",
-		);
-		// Treat empty string as unset so users can supply a new secret
-		const rawMinioPassword =
-			answers.MINIO_ROOT_PASSWORD ?? process.env.MINIO_ROOT_PASSWORD;
-		const existingMinioPassword = rawMinioPassword || undefined;
-		answers.API_MINIO_SECRET_KEY = await promptInput(
-			"API_MINIO_SECRET_KEY",
-			"Minio secret key:",
-			existingMinioPassword ?? "password",
-		);
-		if (existingMinioPassword !== undefined) {
-			// Configured non-empty password found, validate against it
-			const minioPassword = existingMinioPassword;
-			while (answers.API_MINIO_SECRET_KEY !== minioPassword) {
-				console.warn("⚠️ API_MINIO_SECRET_KEY must match MINIO_ROOT_PASSWORD.");
-				answers.API_MINIO_SECRET_KEY = await promptInput(
-					"API_MINIO_SECRET_KEY",
-					"Minio secret key:",
-					minioPassword, // Use configured password as default
-				);
-			}
-			console.log("✅ API_MINIO_SECRET_KEY matches MINIO_ROOT_PASSWORD");
-		} else {
-			// No configured value (or empty): set both answers.MINIO_ROOT_PASSWORD and
-			// process.env.MINIO_ROOT_PASSWORD to answers.API_MINIO_SECRET_KEY
-			// so the chosen API_MINIO_SECRET_KEY becomes the stored Minio password
-			answers.MINIO_ROOT_PASSWORD = answers.API_MINIO_SECRET_KEY;
-			process.env.MINIO_ROOT_PASSWORD = answers.API_MINIO_SECRET_KEY;
-			console.log(
-				"ℹ️  MINIO_ROOT_PASSWORD will be set to match API_MINIO_SECRET_KEY",
-			);
-		}
-		answers.API_MINIO_TEST_END_POINT = await promptInput(
-			"API_MINIO_TEST_END_POINT",
-			"Minio test endpoint:",
-			"minio-test",
-		);
-		answers.API_MINIO_USE_SSL = await promptList(
-			"API_MINIO_USE_SSL",
-			"Use Minio SSL?",
-			["true", "false"],
-			"false",
-		);
-		answers.API_POSTGRES_DATABASE = await promptInput(
-			"API_POSTGRES_DATABASE",
-			"Postgres database:",
-			"talawa",
-		);
-		answers.API_POSTGRES_HOST = await promptInput(
-			"API_POSTGRES_HOST",
-			"Postgres host:",
-			"postgres",
-		);
-		// Treat empty string as unset so users can supply a new secret
-		const rawPostgresPassword =
-			answers.POSTGRES_PASSWORD ?? process.env.POSTGRES_PASSWORD;
-		const postgresPassword = rawPostgresPassword || undefined;
-		answers.API_POSTGRES_PASSWORD = await promptInput(
-			"API_POSTGRES_PASSWORD",
-			"Postgres password:",
-			postgresPassword ?? "password",
-		);
-		if (postgresPassword !== undefined) {
-			// Configured non-empty password found, validate against it
-			const postgresPasswordLocal = postgresPassword;
-			while (answers.API_POSTGRES_PASSWORD !== postgresPasswordLocal) {
-				console.warn("⚠️ API_POSTGRES_PASSWORD must match POSTGRES_PASSWORD.");
-				answers.API_POSTGRES_PASSWORD = await promptInput(
-					"API_POSTGRES_PASSWORD",
-					"Postgres password:",
-					postgresPasswordLocal, // Use configured password as default
-				);
-			}
-			console.log("✅ API_POSTGRES_PASSWORD matches POSTGRES_PASSWORD");
-		} else {
-			// No configured value (or empty): set both answers.POSTGRES_PASSWORD and
-			// process.env.POSTGRES_PASSWORD to answers.API_POSTGRES_PASSWORD
-			// so the chosen API_POSTGRES_PASSWORD becomes the stored Postgres password
-			answers.POSTGRES_PASSWORD = answers.API_POSTGRES_PASSWORD;
-			process.env.POSTGRES_PASSWORD = answers.API_POSTGRES_PASSWORD;
-			console.log(
-				"ℹ️  POSTGRES_PASSWORD will be set to match API_POSTGRES_PASSWORD",
-			);
-		}
-		answers.API_POSTGRES_PORT = await promptInput(
-			"API_POSTGRES_PORT",
-			"Postgres port:",
-			"5432",
-			validatePort,
-		);
-		answers.API_POSTGRES_SSL_MODE = await promptList(
-			"API_POSTGRES_SSL_MODE",
-			"Use Postgres SSL?",
-			["true", "false"],
-			"false",
-		);
-		answers.API_POSTGRES_TEST_HOST = await promptInput(
-			"API_POSTGRES_TEST_HOST",
-			"Postgres test host:",
-			"postgres-test",
-		);
-		answers.API_POSTGRES_USER = await promptInput(
-			"API_POSTGRES_USER",
-			"Postgres user:",
-			"talawa",
-		);
-	} catch (err) {
-		await handlePromptError(err);
-	}
-	return answers;
-}
-export async function cloudbeaverSetup(
-	answers: SetupAnswers,
-): Promise<SetupAnswers> {
-	try {
-		answers.CLOUDBEAVER_ADMIN_NAME = await promptInput(
-			"CLOUDBEAVER_ADMIN_NAME",
-			"CloudBeaver admin name:",
-			"talawa",
-			validateCloudBeaverAdmin,
-		);
-		answers.CLOUDBEAVER_ADMIN_PASSWORD = await promptInput(
-			"CLOUDBEAVER_ADMIN_PASSWORD",
-			"CloudBeaver admin password:",
-			process.env.CLOUDBEAVER_ADMIN_PASSWORD ?? "",
-			validateCloudBeaverPassword,
-		);
-		answers.CLOUDBEAVER_MAPPED_HOST_IP = await promptInput(
-			"CLOUDBEAVER_MAPPED_HOST_IP",
-			"CloudBeaver mapped host IP:",
-			"127.0.0.1",
-		);
-		answers.CLOUDBEAVER_MAPPED_PORT = await promptInput(
-			"CLOUDBEAVER_MAPPED_PORT",
-			"CloudBeaver mapped port:",
-			"8978",
-			validatePort,
-		);
-		answers.CLOUDBEAVER_SERVER_NAME = await promptInput(
-			"CLOUDBEAVER_SERVER_NAME",
-			"CloudBeaver server name:",
-			"Talawa CloudBeaver Server",
-		);
-		answers.CLOUDBEAVER_SERVER_URL = await promptInput(
-			"CLOUDBEAVER_SERVER_URL",
-			"CloudBeaver server URL:",
-			"http://127.0.0.1:8978",
-			validateCloudBeaverURL,
-		);
-	} catch (err) {
-		await handlePromptError(err);
-	}
-	return answers;
-}
-export async function minioSetup(answers: SetupAnswers): Promise<SetupAnswers> {
-	try {
-		answers.MINIO_BROWSER = await promptInput(
-			"MINIO_BROWSER",
-			"Minio browser (on/off):",
-			answers.CI === "true" ? "off" : "on",
-		);
-		if (answers.CI === "false") {
-			answers.MINIO_API_MAPPED_HOST_IP = await promptInput(
-				"MINIO_API_MAPPED_HOST_IP",
-				"Minio API mapped host IP:",
-				"127.0.0.1",
-			);
-			answers.MINIO_API_MAPPED_PORT = await promptInput(
-				"MINIO_API_MAPPED_PORT",
-				"Minio API mapped port:",
-				"9000",
-				validatePort,
-			);
-			answers.MINIO_CONSOLE_MAPPED_HOST_IP = await promptInput(
-				"MINIO_CONSOLE_MAPPED_HOST_IP",
-				"Minio console mapped host IP:",
-				"127.0.0.1",
-			);
-			answers.MINIO_CONSOLE_MAPPED_PORT = await promptInput(
-				"MINIO_CONSOLE_MAPPED_PORT",
-				"Minio console mapped port:",
-				"9001",
-				validatePort,
-			);
-			let portConflict = true;
-			while (portConflict && answers.CI === "false") {
-				if (
-					answers.MINIO_API_MAPPED_PORT === answers.MINIO_CONSOLE_MAPPED_PORT
-				) {
-					console.warn(
-						"⚠️ Port conflict detected: MinIO API and Console ports must be different.",
-					);
-					answers.MINIO_CONSOLE_MAPPED_PORT = await promptInput(
-						"MINIO_CONSOLE_MAPPED_PORT",
-						"Please enter a different Minio console mapped port:",
-						String(Number(answers.MINIO_API_MAPPED_PORT) + 1),
-						validatePort,
-					);
-				} else {
-					portConflict = false;
-				}
-			}
-		}
-		// Use already-synced API_MINIO_SECRET_KEY as default if available
-		const minioPasswordDefault =
-			answers.API_MINIO_SECRET_KEY ??
-			answers.MINIO_ROOT_PASSWORD ??
-			process.env.MINIO_ROOT_PASSWORD ??
-			"password";
-		answers.MINIO_ROOT_PASSWORD = await promptInput(
-			"MINIO_ROOT_PASSWORD",
-			"Minio root password:",
-			minioPasswordDefault,
-		);
-		// Sync back to API_MINIO_SECRET_KEY if it was set
-		if (answers.API_MINIO_SECRET_KEY !== undefined) {
-			if (answers.MINIO_ROOT_PASSWORD !== answers.API_MINIO_SECRET_KEY) {
-				// User changed MINIO_ROOT_PASSWORD, update API_MINIO_SECRET_KEY to match
-				answers.API_MINIO_SECRET_KEY = answers.MINIO_ROOT_PASSWORD;
-				process.env.MINIO_ROOT_PASSWORD = answers.MINIO_ROOT_PASSWORD;
-				console.log(
-					"ℹ️  API_MINIO_SECRET_KEY updated to match MINIO_ROOT_PASSWORD",
-				);
-			}
-		} else {
-			// No API_MINIO_SECRET_KEY set yet, set it now
-			answers.API_MINIO_SECRET_KEY = answers.MINIO_ROOT_PASSWORD;
-			process.env.MINIO_ROOT_PASSWORD = answers.MINIO_ROOT_PASSWORD;
-		}
-		answers.MINIO_ROOT_USER = await promptInput(
-			"MINIO_ROOT_USER",
-			"Minio root user:",
-			"talawa",
-		);
-	} catch (err) {
-		await handlePromptError(err);
-	}
-	return answers;
-}
-export async function postgresSetup(
-	answers: SetupAnswers,
-): Promise<SetupAnswers> {
-	try {
-		answers.POSTGRES_DB = await promptInput(
-			"POSTGRES_DB",
-			"Postgres database:",
-			"talawa",
-		);
-		if (answers.CI === "false") {
-			answers.POSTGRES_MAPPED_HOST_IP = await promptInput(
-				"POSTGRES_MAPPED_HOST_IP",
-				"Postgres mapped host IP:",
-				"127.0.0.1",
-			);
-			answers.POSTGRES_MAPPED_PORT = await promptInput(
-				"POSTGRES_MAPPED_PORT",
-				"Postgres mapped port:",
-				"5432",
-				validatePort,
-			);
-		}
-		// Use already-synced API_POSTGRES_PASSWORD as default if available
-		const postgresPasswordDefault =
-			answers.API_POSTGRES_PASSWORD ??
-			answers.POSTGRES_PASSWORD ??
-			process.env.POSTGRES_PASSWORD ??
-			"password";
-		answers.POSTGRES_PASSWORD = await promptInput(
-			"POSTGRES_PASSWORD",
-			"Postgres password:",
-			postgresPasswordDefault,
-		);
-		// Sync back to API_POSTGRES_PASSWORD if it was set
-		if (answers.API_POSTGRES_PASSWORD !== undefined) {
-			if (answers.POSTGRES_PASSWORD !== answers.API_POSTGRES_PASSWORD) {
-				// User changed POSTGRES_PASSWORD, update API_POSTGRES_PASSWORD to match
-				answers.API_POSTGRES_PASSWORD = answers.POSTGRES_PASSWORD;
-				process.env.POSTGRES_PASSWORD = answers.POSTGRES_PASSWORD;
-				console.log(
-					"ℹ️  API_POSTGRES_PASSWORD updated to match POSTGRES_PASSWORD",
-				);
-			}
-		} else {
-			// No API_POSTGRES_PASSWORD set yet, set it now
-			answers.API_POSTGRES_PASSWORD = answers.POSTGRES_PASSWORD;
-			process.env.POSTGRES_PASSWORD = answers.POSTGRES_PASSWORD;
-		}
-		answers.POSTGRES_USER = await promptInput(
-			"POSTGRES_USER",
-			"Postgres user:",
-			"talawa",
-		);
-	} catch (err) {
-		await handlePromptError(err);
-	}
-	return answers;
-}
-export async function caddySetup(answers: SetupAnswers): Promise<SetupAnswers> {
-	try {
-		answers.CADDY_HTTP_MAPPED_PORT = await promptInput(
-			"CADDY_HTTP_MAPPED_PORT",
-			"Caddy HTTP mapped port:",
-			"80",
-			validatePort,
-		);
-		answers.CADDY_HTTPS_MAPPED_PORT = await promptInput(
-			"CADDY_HTTPS_MAPPED_PORT",
-			"Caddy HTTPS mapped port:",
-			"443",
-			validatePort,
-		);
-		answers.CADDY_HTTP3_MAPPED_PORT = await promptInput(
-			"CADDY_HTTP3_MAPPED_PORT",
-			"Caddy HTTP3 mapped port:",
-			"443",
-			validatePort,
-		);
-		answers.CADDY_TALAWA_API_DOMAIN_NAME = await promptInput(
-			"CADDY_TALAWA_API_DOMAIN_NAME",
-			"Caddy Talawa API domain name:",
-			"localhost",
-		);
-		answers.CADDY_TALAWA_API_EMAIL = await promptInput(
-			"CADDY_TALAWA_API_EMAIL",
-			"Caddy Talawa API email:",
-			"talawa@email.com",
-			validateEmail,
-		);
-		answers.CADDY_TALAWA_API_HOST = await promptInput(
-			"CADDY_TALAWA_API_HOST",
-			"Caddy Talawa API host:",
-			"api",
-		);
-		answers.CADDY_TALAWA_API_PORT = await promptInput(
-			"CADDY_TALAWA_API_PORT",
-			"Caddy Talawa API port:",
-			"4000",
-			validatePort,
-		);
 	} catch (err) {
 		await handlePromptError(err);
 	}
